+++ code/network/multi.h
@@ -122,7 +122,7 @@ class player;
 //
 
 // netplayer management
-#define NET_PLAYER_INDEX(np)	(np-Net_players)
+#define NET_PLAYER_INDEX(np)	((int)(np-Net_players))
 #define NET_PLAYER_NUM(np)		(NET_PLAYER_INDEX(np))
 #define MY_NET_PLAYER_NUM		(NET_PLAYER_INDEX(Net_player))
