+++ code/object/object.h
@@ -231,8 +231,8 @@ extern object *Player_obj;	// Which object is the player. Has to be valid.
 // Use this instead of "objp - Objects" to get an object number
 // given it's pointer.  This way, we can replace it with a macro
 // to check that the pointer is valid for debugging.
-// https://stackoverflow.com/questions/7954439/c-which-character-should-be-used-for-ptrdiff-t-in-printf
-#define OBJ_INDEX(objp) (ptrdiff_t)(objp-Objects)
+// This code will break in 64 bit builds when we have more than 2^31 objects but that will probably never happen
+#define OBJ_INDEX(objp) (int)(objp-Objects)
 
 /*
  *		FUNCTIONS
